pyproject.toml
[project]
name = "gemini-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.0.13",
    "google-genai>=1.21.1",
    "langchain>=0.3.26",
]
